bb-flasher-mspm0: Add I2C support
- Only on linux.
